+++ seqr/views/apis/family_api_tests.py
@@ -68,11 +68,11 @@ class FamilyAPITest(AuthenticationTestCase):
         self.assertSetEqual({PROJECT_GUID}, {i['projectGuid'] for i in response_json['individualsByGuid'].values()})
         self.assertSetEqual({FAMILY_GUID}, {i['familyGuid'] for i in response_json['individualsByGuid'].values()})
 
-        self.assertEqual(len(response_json['samplesByGuid']), 4)
+        self.assertEqual(len(response_json['samplesByGuid']), 6)
         self.assertSetEqual(set(next(iter(response_json['samplesByGuid'].values())).keys()), SAMPLE_FIELDS)
         self.assertSetEqual({PROJECT_GUID}, {s['projectGuid'] for s in response_json['samplesByGuid'].values()})
         self.assertSetEqual({FAMILY_GUID}, {s['familyGuid'] for s in response_json['samplesByGuid'].values()})
-        self.assertEqual(len(individual['sampleGuids']), 2)
+        self.assertEqual(len(individual['sampleGuids']), 4)
         self.assertTrue(set(individual['sampleGuids']).issubset(set(response_json['samplesByGuid'].keys())))
 
         self.assertEqual(len(response_json['igvSamplesByGuid']), 1)
@@ -387,9 +387,9 @@ class FamilyAPITest(AuthenticationTestCase):
             'errors': ['Could not find families with the following previous IDs: 1_old'], 'warnings': []})
 
         # send valid request
-        data = b'Family ID	Previous Family ID	Display Name	Description	Coded Phenotype\n\
-"1_renamed"	"1"	"1"	"family one description"	""\n\
-"2"	""	"2"	"family two description"	""'
+        data = b'Family ID	Previous Family ID	Display Name	Description	Phenotype Description	MONDO ID\n\
+"1_renamed"	"1"	"1"	"family one description"	"dystrophy"	"MONDO:12345"\n\
+"2"	""	"2"	"family two description"	""	""'
 
         response = self.client.post(url, {'f': SimpleUploadedFile("1000_genomes demo_families.tsv", data)})
         self.assertEqual(response.status_code, 200)
@@ -409,6 +409,8 @@ class FamilyAPITest(AuthenticationTestCase):
         family_1 = response_json['familiesByGuid'][FAMILY_GUID]
         self.assertEqual(family_1['description'], 'family one description')
         self.assertEqual(family_1['familyId'], '1_renamed')
+        self.assertEqual(family_1['codedPhenotype'], 'dystrophy')
+        self.assertEqual(family_1['mondoId'], 'MONDO:12345')
         family_2 = response_json['familiesByGuid'][FAMILY_GUID2]
         self.assertEqual(family_2['description'], 'family two description')
         self.assertEqual(family_2['familyId'], '2')