+++ seqr/views/apis/individual_api.py
@@ -754,7 +754,7 @@ def save_individuals_metadata_table_handler(request, project_guid, upload_file_i
 def get_individual_rna_seq_data(request, individual_guid):
     individual = Individual.objects.get(guid=individual_guid)
     check_project_permissions(individual.family.project, request.user)
-    sample = Sample.objects.get(individual=individual, is_active=True, sample_type=Sample.SAMPLE_TYPE_RNA)
+    sample = Sample.objects.get(individual=individual, is_active=True, sample_type=Sample.SAMPLE_TYPE_RNA, tissue_type=None)
 
     rna_seq_data = {
         data['geneId']: data for data in get_json_for_rna_seq_outliers(RnaSeqOutlier.objects.filter(sample=sample))