+++ seqr/views/apis/report_api_tests.py
@@ -1,5 +1,6 @@
 from django.urls.base import reverse
 from django.utils.dateparse import parse_datetime
+import json
 import mock
 import pytz
 import responses
@@ -265,6 +266,53 @@ EXPECTED_SAMPLE_METADATA_ROW = {
     "multiple_datasets": "No",
 }
 EXPECTED_SAMPLE_METADATA_ROW.update(EXPECTED_NO_AIRTABLE_SAMPLE_METADATA_ROW)
+
+MOCK_DATA_MODEL_URL = 'http://raw.githubusercontent.com/gregor_data_model.json'
+MOCK_DATA_MODEL = {
+    'name': 'test data model',
+    'tables': [
+        {
+            'table': 'subject',
+            'required': True,
+            'columns': [{'column': 'subject_id', 'required': True}],
+        },
+        {
+            'table': 'participant',
+            'required': True,
+            'columns': [
+                {'column': 'participant_id', 'required': True},
+                {'column': 'internal_project_id'},
+                {'column': 'gregor_center', 'required': True, 'enumerations': ['BCM', 'BROAD', 'UW']},
+                {'column': 'consent_code', 'required': True, 'enumerations': ['GRU', 'HMB']},
+                {'column': 'recontactable', 'enumerations': ['Yes', 'No', 'Unknown']},
+                {'column': 'prior_testing'},
+                {'column': 'family_id', 'required': True},
+                {'column': 'paternal_id'},
+                {'column': 'maternal_id'},
+                {'column': 'proband_relationship', 'required': True},
+                {'column': 'sex', 'required': True, 'enumerations': ['Male', 'Female', 'Unknown']},
+                {'column': 'reported_race', 'enumerations': ['Asian', 'White', 'Black', 'Unknown']},
+                {'column': 'reported_ethnicity', 'enumerations': ['Hispanic', 'Not Hispanic', 'Unknown']},
+                {'column': 'ancestry_metadata'},
+                {'column': 'affected_status', 'required': True, 'enumerations': ['Affected', 'Unaffected', 'Unknown']},
+                {'column': 'phenotype_description'},
+                {'column': 'age_at_enrollment'},
+            ],
+        },
+        {
+            'table': 'aligned_dna_short_read_set',
+            'columns': [
+                {'column': 'aligned_dna_short_read_set_id', 'required': True},
+                {'column': 'aligned_dna_short_read_id', 'required': True},
+            ],
+        },
+        {
+            'table': 'dna_read_data',
+            'columns': [{'column': 'analyte_id', 'required': True}],
+        },
+    ]
+}
+
 
 def _get_list_param(call, param):
     query_params = call.url.split('?')[1].split('&')
@@ -541,6 +589,18 @@ class ReportAPITest(object):
             -1, "OR(RECORD_ID()='recW24C2CJW5lT64K',RECORD_ID()='reca4hcBnbA2cnZf9')", ['CollaboratorID'])
         self.assertSetEqual({call.request.headers['Authorization'] for call in responses.calls}, {'Bearer mock_key'})
 
+        # Test omit airtable columns
+        responses.reset()
+        response = self.client.get(f'{url}?omitAirtable=true')
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+        self.assertEqual(len(response_json['rows']), 3)
+        expected_samples = {'NA20885', 'NA20888', 'NA20889'}
+        self.assertSetEqual({r['sample_id'] for r in response_json['rows']}, expected_samples)
+        test_row = next(r for r in response_json['rows'] if r['sample_id'] == 'NA20889')
+        self.assertDictEqual(EXPECTED_NO_AIRTABLE_SAMPLE_METADATA_ROW, test_row)
+
         # Test empty project
         empty_project_url = reverse(sample_metadata_export, args=[PROJECT_EMPTY_GUID])
         response = self.client.get(empty_project_url)
@@ -548,7 +608,6 @@ class ReportAPITest(object):
         self.assertDictEqual(response.json(), {'rows': []})
 
         # Test all projects
-        responses.reset()
         all_projects_url = reverse(sample_metadata_export, args=['all'])
         response = self.client.get(all_projects_url)
         self.assertEqual(response.status_code, 200)
@@ -571,39 +630,91 @@ class ReportAPITest(object):
         self.assertEqual(response.status_code, 403)
         self.assertEqual(response.json()['error'], 'Permission Denied')
 
+    @mock.patch('seqr.views.apis.report_api.GREGOR_DATA_MODEL_URL', MOCK_DATA_MODEL_URL)
     @mock.patch('seqr.views.utils.airtable_utils.is_google_authenticated')
     @mock.patch('seqr.views.apis.report_api.datetime')
-    @mock.patch('seqr.views.utils.export_utils.zipfile.ZipFile')
+    @mock.patch('seqr.views.utils.export_utils.open')
+    @mock.patch('seqr.views.utils.export_utils.TemporaryDirectory')
+    @mock.patch('seqr.utils.file_utils.subprocess.Popen')
     @responses.activate
-    def test_gregor_export(self, mock_zip, mock_datetime, mock_google_authenticated):
+    def test_gregor_export(self, mock_subprocess, mock_temp_dir, mock_open, mock_datetime, mock_google_authenticated):
         mock_datetime.now.return_value.year = 2020
         mock_google_authenticated.return_value = False
+        mock_temp_dir.return_value.__enter__.return_value = '/mock/tmp'
+        mock_subprocess.return_value.wait.return_value = 1
+
         responses.add(
             responses.GET, '{}/app3Y97xtbbaOopVR/Samples'.format(AIRTABLE_URL), json=AIRTABLE_GREGOR_SAMPLE_RECORDS,
             status=200)
         responses.add(
             responses.GET, '{}/app3Y97xtbbaOopVR/GREGoR Data Model'.format(AIRTABLE_URL), json=AIRTABLE_GREGOR_RECORDS,
             status=200)
+        responses.add(responses.GET, MOCK_DATA_MODEL_URL, json=MOCK_DATA_MODEL, status=200)
 
-        url = reverse(gregor_export, args=['HMB'])
+        url = reverse(gregor_export)
         self.check_analyst_login(url)
 
-        response = self.client.get(url)
+        response = self.client.post(url, content_type='application/json', data=json.dumps({}))
+        self.assertEqual(response.status_code, 400)
+        self.assertListEqual(response.json()['errors'], ['Missing required field(s): consentCode, deliveryPath'])
+
+        body = {'consentCode': 'HMB', 'deliveryPath': '/test/file'}
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 400)
+        self.assertListEqual(response.json()['errors'], ['Delivery Path must be a valid google bucket path (starts with gs://)'])
+
+        body['deliveryPath'] = 'gs://anvil-upload'
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 400)
+        self.assertListEqual(response.json()['errors'], ['Invalid Delivery Path: folder not found'])
+
+        mock_subprocess.return_value.wait.return_value = 0
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
         self.assertEqual(response.status_code, 403)
         self.assertEqual(response.json()['error'], 'Permission Denied')
 
         mock_google_authenticated.return_value = True
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            response.get('content-disposition'),
-            'attachment; filename="GREGoR Reports HMB.zip"'
-        )
-
-        files = self._get_zip_files(mock_zip, [
-            'participant.tsv', 'family.tsv', 'phenotype.tsv', 'analyte.tsv', 'experiment_dna_short_read.tsv',
-            'aligned_dna_short_read.tsv', 'aligned_dna_short_read_set.tsv', 'called_variants_dna_short_read.tsv',
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 400)
+        expected_files = [
+            'participant', 'family', 'phenotype', 'analyte', 'experiment_dna_short_read',
+            'aligned_dna_short_read', 'aligned_dna_short_read_set', 'called_variants_dna_short_read',
+        ]
+        skipped_file_validation_warnings = [
+            f'No data model found for "{file}" table so no validation was performed' for file in expected_files
+        ]
+        self.assertListEqual(response.json()['warnings'], [
+            'The following tables are required in the data model but absent from the reports: subject',
+            'The following columns are included in the "participant" table but are missing from the data model: age_at_last_observation, ancestry_detail, pmid_id, proband_relationship_detail, sex_detail, twin_id',
+            'The following columns are included in the "participant" data model but are missing in the report: ancestry_metadata',
+            'The following entries are missing recommended "phenotype_description" in the "participant" table: Broad_HG00731, Broad_HG00732, Broad_HG00733, Broad_NA20870, Broad_NA20872, Broad_NA20874, Broad_NA20875, Broad_NA20876, Broad_NA20877, Broad_NA20881',
+            'The following entries are missing recommended "age_at_enrollment" in the "participant" table: Broad_HG00731, Broad_HG00732, Broad_HG00733, Broad_NA19678, Broad_NA19679, Broad_NA20870, Broad_NA20872, Broad_NA20874, Broad_NA20875, Broad_NA20876, Broad_NA20877, Broad_NA20881',
+        ] + skipped_file_validation_warnings[1:6] + skipped_file_validation_warnings[7:])
+        self.assertListEqual(response.json()['errors'], [
+            'The following entries are missing required "proband_relationship" in the "participant" table: Broad_HG00731, Broad_HG00732, Broad_HG00733, Broad_NA19678, Broad_NA19679, Broad_NA20870, Broad_NA20872, Broad_NA20874, Broad_NA20875, Broad_NA20876, Broad_NA20877, Broad_NA20881',
+            'The following entries have invalid values for "reported_race" in the "participant" table. Allowed values: Asian, White, Black, Unknown. Invalid values: Broad_NA19675_1 (Middle Eastern or North African)',
+            'The following entries are missing required "aligned_dna_short_read_set_id" (from Airtable) in the "aligned_dna_short_read_set" table: NA19675_1',
         ])
+
+        responses.add(responses.GET, MOCK_DATA_MODEL_URL, status=404)
+        responses.calls.reset()
+        mock_subprocess.reset_mock()
+        mock_google_authenticated.return_value = True
+        response = self.client.post(url, content_type='application/json', data=json.dumps(body))
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.json(), {
+            'info': ['Successfully validated and uploaded Gregor Report for 9 families'],
+            'warnings': [
+                'Unable to load data model for validation: 404 Client Error: Not Found for url: http://raw.githubusercontent.com/gregor_data_model.json',
+            ] + skipped_file_validation_warnings,
+        })
+
+        self.assertListEqual(
+            mock_open.call_args_list, [mock.call(f'/mock/tmp/{file}.tsv', 'w') for file in expected_files])
+        files = [
+            [row.split('\t') for row in write_call.args[0].split('\n')]
+            for write_call in mock_open.return_value.__enter__.return_value.write.call_args_list
+        ]
         participant_file, family_file, phenotype_file, analyte_file, experiment_file, read_file, read_set_file, called_file = files
 
         self.assertEqual(len(participant_file), 14)
@@ -689,7 +800,7 @@ class ReportAPITest(object):
         ])
 
         # test airtable calls
-        self.assertEqual(len(responses.calls), 3)
+        self.assertEqual(len(responses.calls), 4)
         sample_filter = "OR({CollaboratorSampleID}='HG00731',{CollaboratorSampleID}='HG00732',{CollaboratorSampleID}='HG00733'," \
                         "{CollaboratorSampleID}='NA19675_1',{CollaboratorSampleID}='NA19678',{CollaboratorSampleID}='NA19679'," \
                         "{CollaboratorSampleID}='NA20870',{CollaboratorSampleID}='NA20872',{CollaboratorSampleID}='NA20874'," \
@@ -714,6 +825,16 @@ class ReportAPITest(object):
         ]
         self._assert_expected_airtable_call(2, "OR(SMID='SM-AGHT',SMID='SM-JDBTM')", metadata_fields)
 
+        self.assertEqual(responses.calls[3].request.url, MOCK_DATA_MODEL_URL)
+
+        # test gsutil commands
+        mock_subprocess.assert_has_calls([
+            mock.call('gsutil ls gs://anvil-upload', stdout=-1, stderr=-2, shell=True),
+            mock.call().wait(),
+            mock.call('gsutil mv /mock/tmp/* gs://anvil-upload', stdout=-1, stderr=-2, shell=True),
+            mock.call().wait(),
+        ])
+
         self.check_no_analyst_no_access(url)
 
 
@@ -729,7 +850,7 @@ class LocalReportAPITest(AuthenticationTestCase, ReportAPITest):
             'WGS__MITO': {'non_demo': 1},
             'WES__SV': {'non_demo': 3},
             'WGS__SV': {'non_demo': 1},
-            'RNA__VARIANTS': {'demo': 1, 'non_demo': 1},
+            'RNA__VARIANTS': {'demo': 1, 'non_demo': 3},
         },
     }
 
@@ -746,6 +867,6 @@ class AnvilReportAPITest(AnvilAuthenticationTestCase, ReportAPITest):
             'WGS__MITO': {'internal': 1},
             'WES__SV': {'internal': 3},
             'WGS__SV': {'external': 1},
-            'RNA__VARIANTS': {'internal': 1, 'demo': 1},
+            'RNA__VARIANTS': {'internal': 3, 'demo': 1},
         },
     }