+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## dev
 * Add RnaSeqSpliceOutlier model (REQUIRES DB MIGRATION)
 
+## 2/24/23
+* Updated Gregor sample manifest (REQUIRES DB MIGRATION)
+* Bumps python to 3.9
+
 ## 2/15/23
 * Support sharded VCFs in AnVIL loading